backend/web/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
mysqlclient
mypy==1.14.1
mypy-extensions==1.0.0
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
httpx==0.27.0
types-beautifulsoup4
pandas-stubs
pandas==1.5.3
pymongo==4.11
pymysql==1.1.1
python-dotenv==1.0.1
numpy==1.24.1
scikit-learn==1.2.0
sqlalchemy==1.4.39
transformers==4.28.1
torch==2.1.0